+++ next/src/components/SlippyMap/index.jsx
@@ -1,8 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import {
-  useContext, useEffect, useState, useRef,
-} from 'react'
-import MapGL, { Source, Layer } from 'react-map-gl'
+import { useContext, useState } from 'react'
+import Map, { Source, Layer } from 'react-map-gl'
 
 import { clusterLayer, clusterCountLayer, unclusteredPointLayer } from './layers'
 import { transformMapOptions, transformSourceOptions } from './options'
@@ -11,15 +9,12 @@ import AlbumContext from '../Context'
 
 const MAPBOX_TOKEN = 'pk.eyJ1IjoiZGFuYWN0aXZlIiwiYSI6ImNreHhqdXkwdjcyZnEzMHBmNzhiOWZsc3QifQ.gCRigL866hVF6GNHoGoyRg'
 
-export default function SlippyMap({ items = [{}], centroid }) {
-  const { geo: { zoom: metaZoom } } = useContext(AlbumContext)
+export default function SlippyMap({ items = [{}], centroid, mapRef }) {
+  const meta = useContext(AlbumContext)
+  const metaZoom = meta?.geo?.zoom ?? 10
   const coordinates = centroid?.coordinates ?? []
   const zoom = centroid?.coordinateAccuracy ?? metaZoom
   const [viewport, setViewport] = useState(transformMapOptions({ coordinates, zoom }))
-  const mapRef = useRef(null)
-  useEffect(() => {
-    setViewport(transformMapOptions({ coordinates, zoom }))
-  }, [centroid])
 
   const onClick = (event) => {
     const feature = event.features[0]
@@ -33,10 +28,8 @@ export default function SlippyMap({ items = [{}], centroid }) {
         return
       }
 
-      setViewport({
-        ...viewport,
-        longitude: feature.geometry.coordinates[0],
-        latitude: feature.geometry.coordinates[1],
+      mapRef.current.flyTo({
+        center: feature.geometry.coordinates,
         zoom: clickZoom,
         transitionDuration: 500,
       })
@@ -45,22 +38,29 @@ export default function SlippyMap({ items = [{}], centroid }) {
 
   const geoJsonSource = transformSourceOptions({ items })
   return (
-    <MapGL
-      {...viewport}
-      width="100%"
-      height="100%"
-      mapStyle="mapbox://styles/mapbox/satellite-streets-v10"
-      onViewportChange={setViewport}
-      mapboxApiAccessToken={MAPBOX_TOKEN}
-      interactiveLayerIds={[clusterLayer.id]}
-      onClick={onClick}
-      ref={mapRef}
-    >
-      <Source id="slippyMap" {...geoJsonSource}>
-        <Layer {...clusterLayer} />
-        <Layer {...clusterCountLayer} />
-        <Layer {...unclusteredPointLayer} />
-      </Source>
-    </MapGL>
+    <>
+      <style global jsx>
+        {`
+        .mapboxgl-control-container {
+          display: none;
+        }
+      `}
+      </style>
+      <Map
+        {...viewport}
+        ref={mapRef}
+        mapStyle="mapbox://styles/mapbox/satellite-streets-v11"
+        mapboxAccessToken={MAPBOX_TOKEN}
+        interactiveLayerIds={[clusterLayer.id]}
+        onClick={onClick}
+        onMove={(evt) => setViewport(evt.viewState)}
+      >
+        <Source id="slippyMap" {...geoJsonSource}>
+          <Layer {...clusterLayer} />
+          <Layer {...clusterCountLayer} />
+          <Layer {...unclusteredPointLayer} />
+        </Source>
+      </Map>
+    </>
   )
 }